+++ packages/blip/app/components/chart/stats.js
@@ -235,24 +235,17 @@ class Stats extends React.Component {
       stats.push(stat)
     }
 
-    const cbgSelected = bgSource === 'cbg'
 
     switch (chartType) {
       case 'daily':
         addStat(commonStats.totalInsulin)
         isAutomatedBasalDevice && addStat(commonStats.timeInAuto)
-        cbgSelected && addStat(commonStats.standardDev)
-        break
-
-      case 'trends':
-        addStat(commonStats.standardDev)
         break
 
       case 'patientStatistics':
         addStat(commonStats.totalInsulin)
         addStat(commonStats.averageDailyDose)
         isAutomatedBasalDevice && addStat(commonStats.timeInAuto)
-        cbgSelected && addStat(commonStats.standardDev)
         break
     }