+++ packages/yourloops/components/patient-list/patient-list-header-filters-label.tsx
@@ -58,18 +58,14 @@ export const PatientListHeaderFiltersLabel: FunctionComponent<PatientListHeaderF
   const { t } = useTranslation()
   const { classes } = useStyles()
   const { filters, resetFilters, hasAnyNonPendingFiltersEnabled } = usePatientListContext()
-  const { allPatientsForSelectedTeamCount } = usePatientContext()
+  const { allNonPendingPatientsForSelectedTeamCount } = usePatientContext()
 
   const { filtersLabel } = usePatientListHeaderFiltersLabelHook({
-    allPatientsForSelectedTeamCount,
+    allNonPendingPatientsForSelectedTeamCount,
     patientsDisplayedCount,
     pendingFilterEnabled: filters.pendingEnabled,
     hasAnyNonPendingFiltersEnabled
   })
-
-  if (!filtersLabel) {
-    return null
-  }
 
   return (
     <Box
@@ -83,7 +79,7 @@ export const PatientListHeaderFiltersLabel: FunctionComponent<PatientListHeaderF
       >
         {filtersLabel}
       </Typography>
-      {!filters.pendingEnabled &&
+      {!filters.pendingEnabled && hasAnyNonPendingFiltersEnabled &&
         <>
           <Divider
             orientation="vertical"