+++ packages/yourloops/test/integration/pages/patient-dashboard/patient-dashboard-for-patient.spec.tsx
@@ -25,7 +25,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import { act, screen, waitFor } from '@testing-library/react'
+import { act, screen, waitFor, within } from '@testing-library/react'
 import { renderPage } from '../../utils/render'
 import { mockDataAPI } from '../../mock/data.api.mock'
 import { mockPatientApiForPatients } from '../../mock/patient.api.mock'
@@ -33,19 +33,21 @@ import { mockPatientLogin } from '../../mock/patient-login.mock'
 import { checkMedicalWidgetForPatient } from '../../assert/medical-widget'
 import { mockMedicalFilesAPI } from '../../mock/medical-files.api.mock'
 import TeamAPI from '../../../../lib/team/team.api'
-import { iTeamOne, buildTeamOne, buildTeamTwo } from '../../mock/team.api.mock'
+import { buildPrivateTeam, buildTeamOne, buildTeamTwo, iTeamOne, mySecondTeamId } from '../../mock/team.api.mock'
 import {
   checkJoinTeamDialog,
   checkJoinTeamDialogCancel,
   checkJoinTeamDialogDisplayErrorMessage,
   checkJoinTeamDialogPrivacyCancel
 } from '../../assert/join-team'
-import { monitoredPatientAsTeamMember } from '../../data/patient.api.data'
+import { monitoredPatientAsTeamMember, monitoredPatientId } from '../../data/patient.api.data'
+import MedicalFilesApi from '../../../../lib/medical-files/medical-files.api'
+import { PRIVATE_TEAM_ID } from '../../../../lib/team/team.hook'
 
 describe('Patient dashboard for HCP', () => {
   const monitoredPatientDashboardRoute = '/dashboard'
 
-  beforeAll(() => {
+  beforeEach(() => {
     mockPatientLogin(monitoredPatientAsTeamMember)
     mockPatientApiForPatients()
     mockDataAPI()
@@ -53,17 +55,46 @@ describe('Patient dashboard for HCP', () => {
     jest.spyOn(TeamAPI, 'getTeams').mockResolvedValue([buildTeamOne(), buildTeamTwo()])
     jest.spyOn(TeamAPI, 'joinTeam').mockResolvedValue()
     jest.spyOn(TeamAPI, 'getTeamFromCode').mockResolvedValue(iTeamOne)
+    localStorage.setItem('selectedTeamId', mySecondTeamId)
   })
 
-  it('should display proper header', async () => {
+  it('should display correct components when patient is in some medical teams', async () => {
     const router = renderPage(monitoredPatientDashboardRoute)
     await waitFor(() => {
       expect(router.state.location.pathname).toEqual(monitoredPatientDashboardRoute)
     })
 
     const secondaryHeader = await screen.findByTestId('patient-nav-bar')
+    expect(MedicalFilesApi.getMedicalReports).toHaveBeenCalledWith(monitoredPatientId, null)
     expect(secondaryHeader).toHaveTextContent('DashboardDailyTrendsDownload report')
     await checkMedicalWidgetForPatient()
+    const dashboard = within(await screen.findByTestId('patient-dashboard'))
+    expect(dashboard.getByText('Data calculated on the last 7 days')).toBeVisible()
+    expect(dashboard.getByText('Patient statistics')).toBeVisible()
+    expect(dashboard.getByText('Device Usage')).toBeVisible()
+
+    expect(dashboard.queryByTestId('medical-files-card')).toBeVisible()
+    expect(dashboard.queryByTestId('monitoring-alert-card')).toBeVisible()
+    expect(dashboard.queryByTestId('chat-card')).toBeVisible()
+  })
+
+  it('should render correct components when patient is in no medical teams', async () => {
+    localStorage.setItem('selectedTeamId', PRIVATE_TEAM_ID)
+    jest.spyOn(TeamAPI, 'getTeams').mockResolvedValue([buildPrivateTeam()])
+
+    await act(async () => {
+      renderPage(monitoredPatientDashboardRoute)
+    })
+
+    const dashboard = within(await screen.findByTestId('patient-dashboard'))
+    expect(dashboard.getByText('Data calculated on the last 7 days')).toBeVisible()
+    expect(dashboard.getByText('Patient statistics')).toBeVisible()
+    expect(dashboard.getByText('Device Usage')).toBeVisible()
+
+    expect(dashboard.queryByTestId('remote-monitoring-card')).not.toBeInTheDocument()
+    expect(dashboard.queryByTestId('medical-files-card')).not.toBeInTheDocument()
+    expect(dashboard.queryByTestId('monitoring-alert-card')).not.toBeInTheDocument()
+    expect(dashboard.queryByTestId('chat-card')).not.toBeInTheDocument()
   })
 
   it('should close the dialog after clicking the join care team button with success message', async () => {